requirements.txt
Flask==2.2.3
Flask-Mail==0.9.1
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
pipenv==2023.3.20
pycodestyle==2.10.0
python-dotenv==0.21.1
PyJWT==2.8.0
pymongo==4.6.3
typing_extensions==4.5.0
virtualenv==20.21.0
Werkzeug==2.2.3
